requirements.txt
asgiref==3.8.1
av==13.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
commonmark==0.9.1
crispy-bootstrap5==2024.10
da-vinci==0.4.0
Django==5.1.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-modeltranslation==0.19.10
django-rosetta==0.10.1
django-thumbnails==0.8.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
numpy==2.1.3
pillow==11.0.0
polib==1.2.0
Pygments==2.18.0
requests==2.32.3
rich==13.0.0
setuptools==75.3.0
shortuuid==1.0.13
sqlparse==0.5.1
urllib3==2.2.3
